+++ app/Filament/Resources/ModerationActions/Pages/ListModerationActions.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace App\Filament\Resources\ModerationActions\Pages;
+
+use App\Filament\Resources\ModerationActions\ModerationActionResource;
+use Filament\Resources\Pages\ListRecords;
+
+class ListModerationActions extends ListRecords
+{
+    protected static string $resource = ModerationActionResource::class;
+}