+++ app/Filament/Resources/ModerationActions/Pages/ViewModerationAction.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace App\Filament\Resources\ModerationActions\Pages;
+
+use App\Filament\Resources\ModerationActions\ModerationActionResource;
+use Filament\Resources\Pages\ViewRecord;
+
+class ViewModerationAction extends ViewRecord
+{
+    protected static string $resource = ModerationActionResource::class;
+}